Move addProjectToProjects to logic.js
